+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Improve UX of query text box
   - The query is now auto-applied when changed (with a 500ms debounce), and drops focus on the text box when Enter is pressed
 
+### Fixed
+
+- Don't show request cancellation dialog if the selected request isn't building/loading
+
 ## [2.3.0] - 2024-11-11
 
 ### Added